\name{Lq_fn}
\alias{Lq_fn}
\title{
Lq Function
}
\description{
Defined as Lq(x)=(x^{1-q}-1)/(1-q), where q is a tuning parameter.
}
\usage{
Lq_fn(x, q)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
A vector or a scalar.
}
  \item{q}{
A tuning parameter, 0<q<=1.
}
}
\value{
The function returns (x^{1-q}-1)/(1-q).
}
\author{
Yichen Qin, Carey E. Priebe
Maintainer: Yichen Qin <qinyn@ucmail.uc.edu>
}
