\name{qclust-package}
\alias{qclust-package}
\docType{package}
\title{
Robust Estimation of Gaussian Mixture Models for Model-Based Clustering
}
\description{
Finite Gaussian mixture models fitted by the robust EM algorithm, and robust model-based clustering for multidimensional data.}
\details{
\tabular{ll}{
Package: \tab qclust\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-04-15\cr
License: \tab GPL-2\cr
URL: \tab \url{http://homepages.uc.edu/~qinyn/qclust/} \cr
}
The main function in this package is Qclust().  It provides a robust estimation of Gaussian mixture models and also generates a robust model-based clustering.
}
\author{
Yichen Qin, Carey E. Priebe\cr
Maintainer: Yichen Qin <qinyn@ucmail.uc.edu>\cr
URL: \url{http://homepages.uc.edu/~qinyn/qclust/}
}
